+++ ext/SerdeTOMLExt/SerdeTOMLExt.jl
@@ -1,0 +1,12 @@
+module SerdeTOMLExt
+
+include("Par.jl")
+using .ParToml
+
+include("De.jl")
+using .DeToml
+
+include("Ser.jl")
+using .SerToml
+
+end
